+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Implemented definition for `UpdateMetadata` token operation and token event which allows to update the metadata of a protocol level token.
 - Implemented definition for `TokenAdminRole` which describes the list of supported roles. 
   Implemented `assignAdminRoles` and `revokeAdminRoles` token operations for RBAC too.
 - Implemented `common::from_bytes_complete` that fails if all bytes are not consumed during deserialization.